requirements.txt
# Core Web Framework
flask==3.0.3
werkzeug==3.0.1
python-dotenv==1.0.0
gunicorn==21.2.0

# AI/ML - Local Processing Only (No API required!)
sentence-transformers==2.6.0
faiss-cpu==1.7.4
numpy==1.26.4
scikit-learn==1.5.0

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.2

# Utilities
pandas==2.2.2
tqdm==4.66.4
requests==2.31.0

# Testing (Development only)
pytest==7.4.4

